reorganize project structure
1. remove `bin/`, place executable directly in root
2. rename `out/` to `export/` for readability
3. create `src/` for source files
4. create sub directory for each library (ex. `lib/hashids/`)
