--- bin/README.md
@@ -1,1 +1,0 @@
-Application's executive files (.exe) will be written to this directory.
