+++ compile_run.bat
@@ -1,3 +1,3 @@
-gcc .\lib\hashids.c tests.c -o .\bin\out.exe
-.\bin\out.exe
+gcc .\lib\hashids.c .\src\tests.c -o ggeny.exe
+ggeny.exe
 pause
