+++ lib/hashids/LICENSE.md
