+++ src/README.md
@@ -1,0 +1,1 @@
+Application's source files will be written to this directory.
